Force vertical split to the far right
Before we were only using `wincmd|` which would create the split in the
current window. This was fine when there was only a single vertical
window, but when multiple windows already existed the split would be in
the middle of the worksplace instead of the far right.

By adding `wincmd L`, we push the window all the way to the right. This
may disrepect the `g:side_search_split_pct` setting, but I think that's
better than the side search showing up in the middle.